Delete old test suite
