+++ tests/CMakeLists.txt
@@ -9,27 +9,11 @@
 
 file(GLOB SOURCES test_*.c)
 
-add_executable(caterva_tests main.c ${SOURCES})
-
-# Allow linking with Blosc2 from external apps defining BLOSC_LIB
-target_link_libraries(caterva_tests ${CATERVA_LIB} ${BLOSC_LIB})
-
-
-if(WIN32)
-    # For a rational of this, see comments in https://github.com/bvdberg/ctest/pull/35
-    set_property(TARGET caterva_tests APPEND_STRING PROPERTY LINK_FLAGS " /INCREMENTAL:NO")
-endif()
-
 foreach (source ${SOURCES})
     get_filename_component(target_name ${source} NAME_WE)
     set(target caterva_${target_name})
-    add_executable(${target}  main.c ${target_name}.c)
+    add_executable(${target} ${target_name}.c)
     target_link_libraries(${target} ${CATERVA_LIB} ${BLOSC_LIB})
-
-    if(WIN32)
-        # For a rational of this, see comments in https://github.com/bvdberg/ctest/pull/35
-        set_property(TARGET ${target} APPEND_STRING PROPERTY LINK_FLAGS " /INCREMENTAL:NO")
-    endif()
     add_test(NAME ${target} COMMAND ${target})
     set_tests_properties(${target} PROPERTIES LABELS "caterva")
 endforeach (source)
