+++ tests/test_copy.c
@@ -11,7 +11,7 @@
 
 #include "test_common.h"
 
-static void test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
+static char* test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
                       caterva_storage_backend_t backend, int64_t *chunkshape, int64_t *blockshape,
                       bool enforceframe, char *filename, caterva_storage_backend_t backend2,
                       int64_t *chunkshape2, int64_t *blockshape2, bool enforceframe2, char *filename2) {
@@ -44,7 +44,7 @@ static void test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, in
 
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     /* Create original data */
@@ -53,22 +53,22 @@ static void test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, in
         buffersize *= shape[i];
     }
     uint8_t *buffer = malloc(buffersize);
-    fill_buf(buffer, itemsize, buffersize / itemsize);
+    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
 
     /* Create caterva_array_t with original data */
     caterva_array_t *src;
-    CATERVA_TEST_ERROR(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
+    MU_ASSERT_CATERVA(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
 
     /* Assert the metalayers creation */
     if (storage.backend == CATERVA_STORAGE_BLOSC) {
         if (blosc2_has_metalayer(src->sc, "random") < 0) {
-            CATERVA_TEST_ERROR(CATERVA_ERR_BLOSC_FAILED);
+            MU_ASSERT_CATERVA(CATERVA_ERR_BLOSC_FAILED);
         }
         double *serializeddata;
         uint32_t len;
         blosc2_get_metalayer(src->sc, "random", (uint8_t **) &serializeddata, &len);
         if (*serializeddata != datatoserialize) {
-            CATERVA_TEST_ERROR(CATERVA_ERR_BLOSC_FAILED);
+            MU_ASSERT_CATERVA(CATERVA_ERR_BLOSC_FAILED);
         }
         free(serializeddata);
     }
@@ -88,40 +88,42 @@ static void test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, in
             }
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     caterva_array_t *dest;
-    CATERVA_TEST_ERROR(caterva_array_copy(ctx, src, &storage2, &dest));
+    MU_ASSERT_CATERVA(caterva_array_copy(ctx, src, &storage2, &dest));
 
     uint8_t *buffer_dest = malloc(buffersize);
-    CATERVA_TEST_ERROR(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
+    MU_ASSERT_CATERVA(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
 
     /* Testing */
-    double tol = (itemsize == 4) ? 1e-6 : 1e-15;
-    assert_buf(buffer, buffer_dest, itemsize, src->size, tol);
+    MU_ASSERT_BUFFER(buffer, buffer_dest, buffersize);
 
     /* Free mallocs */
     free(buffer);
     free(buffer_dest);
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &src));
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &dest));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &src));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &dest));
+
+    return 0;
 }
 
-LWTEST_DATA(copy) {
-    caterva_context_t *ctx;
-};
 
-LWTEST_SETUP(copy) {
+caterva_context_t *ctx;
+
+static char* copy_setup() {
     caterva_config_t cfg = CATERVA_CONFIG_DEFAULTS;
-    caterva_context_new(&cfg, &data->ctx);
+    caterva_context_new(&cfg, &ctx);
+    return 0;
 }
 
-LWTEST_TEARDOWN(copy) {
-    caterva_context_free(&data->ctx);
+static char* copy_teardown() {
+    caterva_context_free(&ctx);
+    return 0;
 }
 
-LWTEST_FIXTURE(copy, 1_double_plain_blosc) {
+static char* copy_1_double_plain_blosc() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 1;
     int64_t shape[] = {17};
@@ -138,11 +140,11 @@ LWTEST_FIXTURE(copy, 1_double_plain_blosc) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
               filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
-LWTEST_FIXTURE(copy, 2_double_blosc_blosc) {
+static char* copy_2_double_blosc_blosc() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 2;
     int64_t shape[] = {40, 40};
@@ -159,12 +161,12 @@ LWTEST_FIXTURE(copy, 2_double_blosc_blosc) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
             filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
 
-LWTEST_FIXTURE(copy, 3_double_blosc_plainbuffer) {
+static char* copy_3_double_blosc_plainbuffer() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 3;
     int64_t shape[] = {134, 56, 204};
@@ -181,12 +183,12 @@ LWTEST_FIXTURE(copy, 3_double_blosc_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
             filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
 
-LWTEST_FIXTURE(copy, 4_float_plainbuffer_blosc_frame) {
+static char* copy_4_float_plainbuffer_blosc_frame() {
     uint8_t itemsize = sizeof(float);
     uint8_t ndim = 4;
     int64_t shape[] = {8, 10, 8, 9};
@@ -203,12 +205,12 @@ LWTEST_FIXTURE(copy, 4_float_plainbuffer_blosc_frame) {
     bool enforceframe2 = true;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
             filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
 
-LWTEST_FIXTURE(copy, 5_double_plainbuffer_plainbuffer) {
+static char* copy_5_double_plainbuffer_plainbuffer() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 5;
     int64_t shape[] = {4, 3, 8, 5, 10};
@@ -225,11 +227,11 @@ LWTEST_FIXTURE(copy, 5_double_plainbuffer_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
             filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
-LWTEST_FIXTURE(copy, 6_uint16_blosc_plainbuffer) {
+static char* copy_6_uint16_blosc_plainbuffer() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 6;
     int64_t shape[] = {4, 3, 8, 5, 6, 5};
@@ -246,11 +248,11 @@ LWTEST_FIXTURE(copy, 6_uint16_blosc_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
               filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
-LWTEST_FIXTURE(copy, 7_float_blosc_blosc) {
+static char* copy_7_float_blosc_blosc() {
     uint8_t itemsize = sizeof(float);
     uint8_t ndim = 7;
     int64_t shape[] = {2, 3, 5, 3, 4, 1, 4};
@@ -267,11 +269,11 @@ LWTEST_FIXTURE(copy, 7_float_blosc_blosc) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
               filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
 
-LWTEST_FIXTURE(copy, 8_uint8_plainbuffer_plainbuffer) {
+static char* copy_8_uint8_plainbuffer_plainbuffer() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 8;
     int64_t shape[] = {4, 3, 6, 8, 2, 5, 9, 3};
@@ -288,6 +290,24 @@ LWTEST_FIXTURE(copy, 8_uint8_plainbuffer_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_copy(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
+    return test_copy(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe,
               filename, backend2, chunkshape2, blockshape2, enforceframe2, filename2);
 }
+
+static char* all_tests() {
+    MU_RUN_SETUP(copy_setup)
+
+    MU_RUN_TEST(copy_1_double_plain_blosc)
+    MU_RUN_TEST(copy_2_double_blosc_blosc)
+    MU_RUN_TEST(copy_3_double_blosc_plainbuffer)
+    MU_RUN_TEST(copy_4_float_plainbuffer_blosc_frame)
+    MU_RUN_TEST(copy_5_double_plainbuffer_plainbuffer)
+    MU_RUN_TEST(copy_6_uint16_blosc_plainbuffer)
+    MU_RUN_TEST(copy_7_float_blosc_blosc)
+    MU_RUN_TEST(copy_8_uint8_plainbuffer_plainbuffer)
+
+    MU_RUN_TEARDOWN(copy_teardown)
+    return 0;
+}
+
+MU_RUN_SUITE("COPY");
