+++ tests/test_persistency.c
@@ -18,7 +18,7 @@
 #define FILE_EXISTS(filename) _access(filename, 0)
 #endif
 
-static void test_persistency(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
+static char* test_persistency(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
                              caterva_storage_backend_t backend, int64_t *chunkshape, int64_t *blockshape,
                              bool enforceframe, char *filename) {
 
@@ -46,7 +46,7 @@ static void test_persistency(caterva_context_t *ctx, uint8_t itemsize, uint8_t n
             }
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     /* Create original data */
@@ -55,51 +55,50 @@ static void test_persistency(caterva_context_t *ctx, uint8_t itemsize, uint8_t n
         buffersize *= shape[i];
     }
     uint8_t *buffer = malloc(buffersize);
-    fill_buf(buffer, itemsize, buffersize / itemsize);
+    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
 
     /* Create caterva_array_t with original data */
     caterva_array_t *src;
-    CATERVA_TEST_ERROR(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
+    MU_ASSERT_CATERVA(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
 
     caterva_array_t *dest;
     caterva_array_from_file(ctx, filename, true, &dest);
 
     /* Fill dest array with caterva_array_t data */
     uint8_t *buffer_dest = malloc(buffersize);
-    CATERVA_TEST_ERROR(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
+    MU_ASSERT_CATERVA(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
 
     /* Testing */
-    double tol = (itemsize == 4) ? 1e-6 : 1e-15;
-    assert_buf(buffer, buffer_dest, itemsize, dest->size, tol);
+    MU_ASSERT_BUFFER(buffer, buffer_dest, buffersize);
 
     /* Free mallocs */
     free(buffer);
     free(buffer_dest);
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &src));
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &dest));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &src));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &dest));
     if (FILE_EXISTS(filename) != -1) {
         remove(filename);
     }
+    return 0;
 }
 
+caterva_context_t *ctx;
+char *filename;
 
-LWTEST_DATA(persistency) {
-    caterva_context_t *ctx;
-    char *filename;
-};
-
-LWTEST_SETUP(persistency) {
+static char* persistency_setup() {
     caterva_config_t cfg = CATERVA_CONFIG_DEFAULTS;
-    caterva_context_new(&cfg, &data->ctx);
-    data->filename = "persistency.caterva";
+    caterva_context_new(&cfg, &ctx);
+    filename = "persistency.caterva";
+    return 0;
 }
 
-LWTEST_TEARDOWN(persistency) {
-    caterva_context_free(&data->ctx);
+static char* persistency_teardown() {
+    caterva_context_free(&ctx);
+    return 0;
 }
 
 
-LWTEST_FIXTURE(persistency, 1_uint16) {
+static  char* persistency_1_uint16() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 1;
     int64_t shape[] = {500};
@@ -108,12 +107,11 @@ LWTEST_FIXTURE(persistency, 1_uint16) {
     int64_t chunkshape[] = {200};
     int64_t blockshape[] = {80};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(persistency, 2_uint8) {
+static  char* persistency_2_uint8() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 2;
     int64_t shape[] = {400, 300};
@@ -122,13 +120,11 @@ LWTEST_FIXTURE(persistency, 2_uint8) {
     int64_t chunkshape[] = {55, 67};
     int64_t blockshape[] = {40, 40};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-
-LWTEST_FIXTURE(persistency, 3_double) {
+static  char* persistency_3_double() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 3;
     int64_t shape[] = {134, 56, 204};
@@ -137,14 +133,11 @@ LWTEST_FIXTURE(persistency, 3_double) {
     int64_t chunkshape[] = {26, 17, 34};
     int64_t blockshape[] = {11, 8, 13};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
-
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-
-LWTEST_FIXTURE(persistency, 4_float) {
+static  char* persistency_4_float() {
     uint8_t itemsize = sizeof(float);
     uint8_t ndim = 4;
     int64_t shape[] = {10, 13, 18, 25};
@@ -153,12 +146,11 @@ LWTEST_FIXTURE(persistency, 4_float) {
     int64_t chunkshape[] = {9, 5, 11, 10};
     int64_t blockshape[] = {2, 2, 5, 3};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(persistency, 4_uint8) {
+static  char* persistency_4_uint8() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 4;
     int64_t shape[] = {78, 85, 34, 56};
@@ -167,13 +159,12 @@ LWTEST_FIXTURE(persistency, 4_uint8) {
     int64_t chunkshape[] = {23, 12, 24, 50};
     int64_t blockshape[] = {11, 5, 7, 13};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
 
-LWTEST_FIXTURE(persistency, 5_double) {
+static  char* persistency_5_double() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 5;
     int64_t shape[] = {25, 27, 24, 25, 12};
@@ -182,12 +173,11 @@ LWTEST_FIXTURE(persistency, 5_double) {
     int64_t chunkshape[] = {10, 15, 8, 11, 11};
     int64_t blockshape[] = {3, 3, 3, 3, 3};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(persistency, 6_uint16) {
+static  char* persistency_6_uint16() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 6;
     int64_t shape[] = {4, 3, 8, 5, 10, 12};
@@ -196,12 +186,11 @@ LWTEST_FIXTURE(persistency, 6_uint16) {
     int64_t chunkshape[] = {2, 2, 3, 3, 4, 5};
     int64_t blockshape[] = {2, 1, 2, 2, 3, 4};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(persistency, 7_uint64) {
+static  char* persistency_7_uint64() {
     uint8_t itemsize = sizeof(uint64_t);
     uint8_t ndim = 7;
     int64_t shape[] =  {4, 15, 11, 6, 12, 8, 7};
@@ -210,12 +199,11 @@ LWTEST_FIXTURE(persistency, 7_uint64) {
     int64_t chunkshape[] = {4, 3, 4, 3, 3, 5, 3};
     int64_t blockshape[] = {1, 3, 2, 2, 1, 2, 2};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(persistency, 8_uint8) {
+static  char* persistency_8_uint8() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 8;
     int64_t shape[] = {4, 3, 8, 5, 10, 12, 6, 4};
@@ -224,7 +212,25 @@ LWTEST_FIXTURE(persistency, 8_uint8) {
     int64_t chunkshape[] = {3, 2, 3, 3, 4, 5, 4, 2};
     int64_t blockshape[] = {2, 1, 2, 2, 3, 4, 3, 1};
     bool enforceframe = true;
-    char *filename = data->filename;
 
-    test_persistency(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_persistency(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
+
+static char* all_tests() {
+    MU_RUN_SETUP(persistency_setup)
+
+    MU_RUN_TEST(persistency_1_uint16)
+    MU_RUN_TEST(persistency_2_uint8)
+    MU_RUN_TEST(persistency_3_double)
+    MU_RUN_TEST(persistency_4_float)
+    MU_RUN_TEST(persistency_4_uint8)
+    MU_RUN_TEST(persistency_5_double)
+    MU_RUN_TEST(persistency_6_uint16)
+    MU_RUN_TEST(persistency_7_uint64)
+    MU_RUN_TEST(persistency_8_uint8)
+
+    MU_RUN_TEARDOWN(persistency_teardown)
+    return 0;
+}
+
+MU_RUN_SUITE("PERSISTENCY")
