+++ tests/test_roundtrip.c
@@ -11,7 +11,7 @@
 
 #include "test_common.h"
 
-static void test_roundtrip(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
+static char* test_roundtrip(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
                            caterva_storage_backend_t backend, int64_t *chunkshape, int64_t *blockshape,
                            bool enforceframe, char *filename) {
 
@@ -36,7 +36,7 @@ static void test_roundtrip(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndi
             }
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     /* Create original data */
@@ -45,42 +45,44 @@ static void test_roundtrip(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndi
         buffersize *= shape[i];
     }
     uint8_t *buffer = malloc(buffersize);
-    fill_buf(buffer, itemsize, buffersize / itemsize);
+    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
 
     /* Create caterva_array_t with original data */
     caterva_array_t *src;
-    CATERVA_TEST_ERROR(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
+    MU_ASSERT_CATERVA(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
 
     /* Fill dest array with caterva_array_t data */
     uint8_t *buffer_dest = malloc(buffersize);
-    CATERVA_TEST_ERROR(caterva_array_to_buffer(ctx, src, buffer_dest, buffersize));
+    MU_ASSERT_CATERVA(caterva_array_to_buffer(ctx, src, buffer_dest, buffersize));
 
     /* Testing */
-    double tol = (itemsize == 4) ? 1e-6 : 1e-15;
-    assert_buf(buffer, buffer_dest, itemsize, src->size, tol);
+    MU_ASSERT_BUFFER(buffer, buffer_dest, buffersize);
 
     /* Free mallocs */
     free(buffer);
     free(buffer_dest);
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &src));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &src));
+    return 0;
 }
 
 
-LWTEST_DATA(roundtrip) {
-    caterva_context_t *ctx;
-};
+caterva_context_t *ctx;
 
-LWTEST_SETUP(roundtrip) {
+static char* roundtrip_setup() {
     caterva_config_t cfg = CATERVA_CONFIG_DEFAULTS;
-    caterva_context_new(&cfg, &data->ctx);
+    cfg.nthreads = 2;
+    cfg.compcodec = BLOSC_BLOSCLZ;
+    caterva_context_new(&cfg, &ctx);
+    return 0;
 }
 
-LWTEST_TEARDOWN(roundtrip) {
-    caterva_context_free(&data->ctx);
+static char* roundtrip_teardown() {
+    caterva_context_free(&ctx);
+    return 0;
 }
 
 
-LWTEST_FIXTURE(roundtrip, 1_uint16_plainbuffer) {
+static char* roundtrip_1_uint16_plainbuffer() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 1;
     int64_t shape[] = {7};
@@ -91,10 +93,10 @@ LWTEST_FIXTURE(roundtrip, 1_uint16_plainbuffer) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(roundtrip, 2_uint8_plainbuffer) {
+static char* roundtrip_2_uint8_plainbuffer() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 2;
     int64_t shape[] = {4, 3};
@@ -105,11 +107,11 @@ LWTEST_FIXTURE(roundtrip, 2_uint8_plainbuffer) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
 
-LWTEST_FIXTURE(roundtrip, 3_double_blosc) {
+static char* roundtrip_3_double_blosc() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 3;
     int64_t shape[] = {134, 56, 204};
@@ -120,12 +122,12 @@ LWTEST_FIXTURE(roundtrip, 3_double_blosc) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 
 }
 
 
-LWTEST_FIXTURE(roundtrip, 4_float_blosc) {
+static char* roundtrip_4_float_blosc() {
     uint8_t itemsize = sizeof(float);
     uint8_t ndim = 4;
     int64_t shape[] = {4, 3, 8, 5};
@@ -136,10 +138,10 @@ LWTEST_FIXTURE(roundtrip, 4_float_blosc) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(roundtrip, 4_uint8_plainbuffer) {
+static char* roundtrip_4_uint8_plainbuffer() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 4;
     int64_t shape[] = {78, 85, 34, 56};
@@ -150,10 +152,10 @@ LWTEST_FIXTURE(roundtrip, 4_uint8_plainbuffer) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(roundtrip, 5_double_plainbuffer) {
+static char* roundtrip_5_double_plainbuffer() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 5;
     int64_t shape[] = {35, 55, 24, 36, 12};
@@ -164,10 +166,10 @@ LWTEST_FIXTURE(roundtrip, 5_double_plainbuffer) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(roundtrip, 6_uint16_blosc) {
+static char* roundtrip_6_uint16_blosc() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 6;
     int64_t shape[] = {4, 3, 8, 5, 10, 12};
@@ -178,10 +180,10 @@ LWTEST_FIXTURE(roundtrip, 6_uint16_blosc) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(roundtrip, 7_double_plainbuffer) {
+static char* roundtrip_7_double_plainbuffer() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 7;
     int64_t shape[] =  {4, 15, 11, 6, 12, 8, 7};
@@ -192,10 +194,10 @@ LWTEST_FIXTURE(roundtrip, 7_double_plainbuffer) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(roundtrip, 8_uint8_blosc) {
+static char* roundtrip_8_uint8_blosc() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 8;
     int64_t shape[] = {4, 3, 8, 5, 10, 12, 6, 4};
@@ -206,5 +208,25 @@ LWTEST_FIXTURE(roundtrip, 8_uint8_blosc) {
     bool enforceframe = false;
     char *filename = NULL;
 
-    test_roundtrip(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_roundtrip(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
+
+
+static char* all_tests() {
+    MU_RUN_SETUP(roundtrip_setup)
+
+    MU_RUN_TEST(roundtrip_1_uint16_plainbuffer)
+    MU_RUN_TEST(roundtrip_2_uint8_plainbuffer)
+    MU_RUN_TEST(roundtrip_3_double_blosc)
+    MU_RUN_TEST(roundtrip_4_float_blosc)
+    MU_RUN_TEST(roundtrip_4_uint8_plainbuffer)
+    MU_RUN_TEST(roundtrip_5_double_plainbuffer)
+    MU_RUN_TEST(roundtrip_6_uint16_blosc)
+    MU_RUN_TEST(roundtrip_7_double_plainbuffer)
+    MU_RUN_TEST(roundtrip_8_uint8_blosc)
+
+    MU_RUN_TEARDOWN(roundtrip_teardown)
+    return 0;
+}
+
+MU_RUN_SUITE("ROUNDTRIP")
