+++ tests/test_serialize.c
@@ -11,7 +11,7 @@
 
 #include "test_common.h"
 
-static void test_serialize(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
+static char* test_serialize(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
                            caterva_storage_backend_t backend, int64_t *chunkshape, int64_t *blockshape,
                            bool enforceframe, char *filename) {
 
@@ -36,7 +36,7 @@ static void test_serialize(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndi
             }
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     /* Create original data */
@@ -45,11 +45,11 @@ static void test_serialize(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndi
         buffersize *= shape[i];
     }
     uint8_t *buffer = malloc(buffersize);
-    fill_buf(buffer, itemsize, buffersize / itemsize);
+    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
 
     /* Create caterva_array_t with original data */
     caterva_array_t *src;
-    CATERVA_TEST_ERROR(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
+    MU_ASSERT_CATERVA(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
 
     uint8_t *sframe = src->sc->frame->sdata;
     int64_t slen = src->sc->frame->len;
@@ -59,35 +59,35 @@ static void test_serialize(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndi
 
     /* Fill dest array with caterva_array_t data */
     uint8_t *buffer_dest = malloc(buffersize);
-    CATERVA_TEST_ERROR(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
+    MU_ASSERT_CATERVA(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
 
     /* Testing */
-    double tol = (itemsize == 4) ? 1e-6 : 1e-15;
-    assert_buf(buffer, buffer_dest, itemsize, dest->size, tol);
+    MU_ASSERT_BUFFER(buffer, buffer_dest, buffersize);
 
     /* Free mallocs */
     free(buffer);
     free(buffer_dest);
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &src));
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &dest));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &src));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &dest));
+    return 0;
 }
 
 
-LWTEST_DATA(serialize) {
-    caterva_context_t *ctx;
-};
+caterva_context_t *ctx;
 
-LWTEST_SETUP(serialize) {
+static char* serialize_setup() {
     caterva_config_t cfg = CATERVA_CONFIG_DEFAULTS;
-    caterva_context_new(&cfg, &data->ctx);
+    caterva_context_new(&cfg, &ctx);
+    return 0;
 }
 
-LWTEST_TEARDOWN(serialize) {
-    caterva_context_free(&data->ctx);
+static char* serialize_teardown() {
+    caterva_context_free(&ctx);
+    return 0;
 }
 
 
-LWTEST_FIXTURE(serialize, 1_uint16) {
+static char* serialize_1_uint16() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 1;
     int64_t shape[] = {500};
@@ -98,10 +98,10 @@ LWTEST_FIXTURE(serialize, 1_uint16) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(serialize, 2_uint8) {
+static char* serialize_2_uint8() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 2;
     int64_t shape[] = {400, 300};
@@ -112,11 +112,11 @@ LWTEST_FIXTURE(serialize, 2_uint8) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
 
-LWTEST_FIXTURE(serialize, 3_double) {
+static char* serialize_3_double() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 3;
     int64_t shape[] = {134, 56, 204};
@@ -127,12 +127,12 @@ LWTEST_FIXTURE(serialize, 3_double) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 
 }
 
 
-LWTEST_FIXTURE(serialize, 4_float) {
+static char* serialize_4_float() {
     uint8_t itemsize = sizeof(float);
     uint8_t ndim = 4;
     int64_t shape[] = {10, 13, 18, 25};
@@ -143,10 +143,10 @@ LWTEST_FIXTURE(serialize, 4_float) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(serialize, 4_uint8) {
+static char* serialize_4_uint8() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 4;
     int64_t shape[] = {78, 85, 34, 56};
@@ -157,11 +157,11 @@ LWTEST_FIXTURE(serialize, 4_uint8) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
 
-LWTEST_FIXTURE(serialize, 5_double) {
+static char* serialize_5_double() {
     uint8_t itemsize = sizeof(double);
     uint8_t ndim = 5;
     int64_t shape[] = {35, 55, 24, 36, 12};
@@ -172,10 +172,10 @@ LWTEST_FIXTURE(serialize, 5_double) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(serialize, 6_uint16) {
+static char* serialize_6_uint16() {
     uint8_t itemsize = sizeof(uint16_t);
     uint8_t ndim = 6;
     int64_t shape[] = {4, 3, 8, 5, 10, 12};
@@ -186,10 +186,10 @@ LWTEST_FIXTURE(serialize, 6_uint16) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(serialize, 7_uint64) {
+static char* serialize_7_uint64() {
     uint8_t itemsize = sizeof(uint64_t);
     uint8_t ndim = 7;
     int64_t shape[] =  {4, 15, 11, 6, 12, 8, 7};
@@ -200,10 +200,10 @@ LWTEST_FIXTURE(serialize, 7_uint64) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
 
-LWTEST_FIXTURE(serialize, 8_uint8) {
+static char* serialize_8_uint8() {
     uint8_t itemsize = sizeof(uint8_t);
     uint8_t ndim = 8;
     int64_t shape[] = {4, 3, 8, 5, 10, 12, 6, 4};
@@ -214,5 +214,25 @@ LWTEST_FIXTURE(serialize, 8_uint8) {
     bool enforceframe = true;
     char *filename = NULL;
 
-    test_serialize(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
+    return test_serialize(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename);
 }
+
+static char* all_tests() {
+    MU_RUN_SETUP(serialize_setup)
+
+    MU_RUN_TEST(serialize_1_uint16)
+    MU_RUN_TEST(serialize_2_uint8)
+    MU_RUN_TEST(serialize_3_double)
+    MU_RUN_TEST(serialize_4_float)
+    MU_RUN_TEST(serialize_4_uint8)
+    MU_RUN_TEST(serialize_5_double)
+    MU_RUN_TEST(serialize_6_uint16)
+    MU_RUN_TEST(serialize_7_uint64)
+    MU_RUN_TEST(serialize_8_uint8)
+
+
+    MU_RUN_TEARDOWN(serialize_teardown)
+    return 0;
+}
+
+MU_RUN_SUITE("SERIALIZE")
