+++ tests/test_squeeze.c
@@ -11,12 +11,10 @@
 
 #include "test_common.h"
 
-static void test_squeeze(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
-                         caterva_storage_backend_t backend, int64_t *chunkshape, int64_t *blockshape, bool enforceframe, 
-                         char *filename, caterva_storage_backend_t backend2, int64_t *chunkshape2, int64_t *blockshape2,
-                         bool enforceframe2, char *filename2, int64_t *start, int64_t *stop) {
-
-
+static char* test_squeeze(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, int64_t *shape,
+                          caterva_storage_backend_t backend, int64_t *chunkshape, int64_t *blockshape, bool enforceframe, 
+                          char *filename, caterva_storage_backend_t backend2, int64_t *chunkshape2, int64_t *blockshape2,
+                          bool enforceframe2, char *filename2, int64_t *start, int64_t *stop) {
     caterva_params_t params;
     params.itemsize = itemsize;
     params.ndim = ndim;
@@ -38,7 +36,7 @@ static void test_squeeze(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim,
             }
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     /* Create original data */
@@ -47,11 +45,11 @@ static void test_squeeze(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim,
         buffersize *= shape[i];
     }
     uint8_t *buffer = malloc(buffersize);
-    fill_buf(buffer, itemsize, buffersize / itemsize);
+    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
 
     /* Create caterva_array_t with original data */
     caterva_array_t *src;
-    CATERVA_TEST_ERROR(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
+    MU_ASSERT_CATERVA(caterva_array_from_buffer(ctx, buffer, buffersize, &params, &storage, &src));
 
 
     /* Create storage for dest container */
@@ -70,36 +68,39 @@ static void test_squeeze(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim,
             }
             break;
         default:
-            CATERVA_TEST_ERROR(CATERVA_ERR_INVALID_STORAGE);
+            MU_ASSERT_CATERVA(CATERVA_ERR_INVALID_STORAGE);
     }
 
     caterva_array_t *dest;
-    CATERVA_TEST_ERROR(caterva_array_get_slice(ctx, src, start, stop, &storage2, &dest));
+    MU_ASSERT_CATERVA(caterva_array_get_slice(ctx, src, start, stop, &storage2, &dest));
 
-    CATERVA_TEST_ERROR(caterva_array_squeeze(ctx, dest));
+    MU_ASSERT_CATERVA(caterva_array_squeeze(ctx, dest));
 
-    LWTEST_ASSERT_TRUE(src->ndim != dest->ndim);
+    MU_ASSERT_NOT_EQUAL(src->ndim ,dest->ndim);
 
     free(buffer);
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &src));
-    CATERVA_TEST_ERROR(caterva_array_free(ctx, &dest));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &src));
+    MU_ASSERT_CATERVA(caterva_array_free(ctx, &dest));
+    
+    return 0;
 }
 
-LWTEST_DATA(squeeze) {
-    caterva_context_t *ctx;
-};
 
-LWTEST_SETUP(squeeze) {
+caterva_context_t *ctx;
+
+static char* squeeze_setup() {
     caterva_config_t cfg = CATERVA_CONFIG_DEFAULTS;
-    caterva_context_new(&cfg, &data->ctx);
+    caterva_context_new(&cfg, &ctx);
+    return 0;
 }
 
-LWTEST_TEARDOWN(squeeze) {
-    caterva_context_free(&data->ctx);
+static char* squeeze_teardown() {
+    caterva_context_free(&ctx);
+    return 0;
 }
 
 
-LWTEST_FIXTURE(squeeze, 2_float_blosc_plainbuffer) {
+static char* squeeze_2_float_blosc_plainbuffer() {
     int64_t start[] = {5, 20};
     int64_t stop[] = {23, 21};
 
@@ -119,11 +120,11 @@ LWTEST_FIXTURE(squeeze, 2_float_blosc_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
                  backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
 
-LWTEST_FIXTURE(squeeze, 3_float_blosc_blosc) {
+static char* squeeze_3_float_blosc_blosc() {
     int64_t start[] = {5, 20, 60};
     int64_t stop[] = {23, 21, 99};
 
@@ -143,11 +144,11 @@ LWTEST_FIXTURE(squeeze, 3_float_blosc_blosc) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
         backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
 
-LWTEST_FIXTURE(squeeze, 4_double_plainbuffer_blosc) {
+static char* squeeze_4_double_plainbuffer_blosc() {
     int64_t start[] = {5, 20, 10, 10};
     int64_t stop[] = {23, 21, 33, 39};
 
@@ -167,12 +168,12 @@ LWTEST_FIXTURE(squeeze, 4_double_plainbuffer_blosc) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
                  backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
 
 
-LWTEST_FIXTURE(squeeze, 5_uint8_plainbuffer_plainbuffer) {
+static char* squeeze_5_uint8_plainbuffer_plainbuffer() {
     int64_t start[] = {1, 12, 3, 12, 6};
     int64_t stop[] = {16, 21, 19, 13, 21};
 
@@ -192,11 +193,11 @@ LWTEST_FIXTURE(squeeze, 5_uint8_plainbuffer_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
                  backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
 
-LWTEST_FIXTURE(squeeze, 6_float_blosc_plainbuffer_frame) {
+static char* squeeze_6_float_blosc_plainbuffer_frame() {
     int64_t start[] = {3, 3, 2, 1, 0, 4};
     int64_t stop[] = {8, 10, 5, 2, 1, 9};
 
@@ -216,11 +217,11 @@ LWTEST_FIXTURE(squeeze, 6_float_blosc_plainbuffer_frame) {
     bool enforceframe2 = true;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
                  backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
 
-LWTEST_FIXTURE(squeeze, 7_uint64_blosc_frame_plainbuffer) {
+static char* squeeze_7_uint64_blosc_frame_plainbuffer() {
     int64_t start[] = {5, 3, 3, 2, 1, 0, 4};
     int64_t stop[] = {6, 8, 10, 5, 2, 1, 9};
 
@@ -240,11 +241,11 @@ LWTEST_FIXTURE(squeeze, 7_uint64_blosc_frame_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
                  backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
 
-LWTEST_FIXTURE(squeeze, 8_uint16_blosc_frame_plainbuffer) {
+static char* squeeze_8_uint16_blosc_frame_plainbuffer() {
     int64_t start[] = {5, 4, 3, 1, 2, 1, 0, 3};
     int64_t stop[] = {6, 8, 7, 4, 7, 2, 1, 9};
 
@@ -264,6 +265,23 @@ LWTEST_FIXTURE(squeeze, 8_uint16_blosc_frame_plainbuffer) {
     bool enforceframe2 = false;
     char *filename2 = NULL;
 
-    test_squeeze(data->ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
+    return test_squeeze(ctx, itemsize, ndim, shape, backend, chunkshape, blockshape, enforceframe, filename,
                  backend2, chunkshape2, blockshape2, enforceframe2, filename2, start, stop);
 }
+
+static char* all_tests() {
+    MU_RUN_SETUP(squeeze_setup);
+
+    MU_RUN_TEST(squeeze_2_float_blosc_plainbuffer);
+    MU_RUN_TEST(squeeze_3_float_blosc_blosc);
+    MU_RUN_TEST(squeeze_4_double_plainbuffer_blosc);
+    MU_RUN_TEST(squeeze_5_uint8_plainbuffer_plainbuffer);
+    MU_RUN_TEST(squeeze_6_float_blosc_plainbuffer_frame);
+    MU_RUN_TEST(squeeze_7_uint64_blosc_frame_plainbuffer);
+    MU_RUN_TEST(squeeze_8_uint16_blosc_frame_plainbuffer);
+
+    MU_RUN_TEARDOWN(squeeze_teardown);
+    return 0;
+}
+
+MU_RUN_SUITE("SQUEEZE")
